+++ saleor/graphql/order/tests/queries/test_order_by_token.py
@@ -581,10 +581,6 @@ QUERY_ORDER_BY_TOKEN_WITH_PAYMENT = """
               created
               modified
               paymentMethodType
-              transactions{
-                gatewayResponse
-
-              }
               actions
               capturedAmount{
                 amount